- Remove DataLoader support for SqueezeNet. Squeezenet's data set was in a different format, and the code becomes tricky if we want to support both BERT data and SqueezeNet data in one class. As we don't have plan to support SqueezeNet in short term, I have disabled the build for SqueezeNet.

- BertLoss was expecting LM_label in [batch_size, max_sequence_length (512)], however, the real data comes in [batch_size, num_prediction_per_sequence (80)]. To accommodate the difference, I added a few nodes to scatter the input to expected shape. This is not the optimal solution. Ideally, when we have cuda kernel for NDGather, we should first nd_gather the prediction, before feeding into the loss function.

- Data Loader supports load X files at a time. At the end of each file, it release memory of current file and load data from the next file. TODO: handle the release and load in another thread.

- Disabled int32_t index type support for SparseSoftmaxCrossEntropy. Clearly, kernel registry doesn't support two typed template class

- DISABLED_TestStatusPropagation: TO FIX

+++ onnxruntime/contrib_ops/cpu/cross_entropy.cc
@@ -184,7 +184,7 @@ Status SparseSoftmaxCrossEntropy<T>::Compute(OpKernelContext* context) const {
   Tensor* loss = context->Output(0, output_shape);
 
   const float* logit_data = logit.template Data<float>();
-  const int* label_data = label.template Data<int>();
+  const int64_t* label_data = label.template Data<int64_t>();
   float* loss_data = loss->template MutableData<float>();
 
   // computation begins here
@@ -256,7 +256,7 @@ Status SparseSoftmaxCrossEntropyGrad<T>::Compute(OpKernelContext* context) const
 
   const float* dY_data = dY.template Data<float>();
   const float* logit_data = logit.template Data<float>();
-  const int* label_data = label.template Data<int>();
+  const int64_t* label_data = label.template Data<int64_t>();
   float* d_logits_data = d_logits->template MutableData<float>();
 
   // computation begins here
@@ -284,7 +284,7 @@ Status SparseSoftmaxCrossEntropyGrad<T>::Compute(OpKernelContext* context) const
     ORT_ENFORCE(weight_shape == label_shape, "The shape in weights and labels is different");
     const float* weight_data = weight.template Data<float>();
     for (int i = 0; i < n; i++) {
-      int label_sample = label_data[i];
+      int64_t label_sample = label_data[i];
       float weight_smaple = weight_data[i] * (*dY_data);
       for (int j = 0; j < d; j++) {
         int index = i * d + j;
@@ -293,7 +293,7 @@ Status SparseSoftmaxCrossEntropyGrad<T>::Compute(OpKernelContext* context) const
     }
   } else {
     for (int i = 0; i < n; i++) {
-      int idx = i * d + label_data[i];
+      int64_t idx = i * d + label_data[i];
       prob[idx] = prob[idx] - (float)1;
     }
     // d_logits = dY * backprop, dY is a scalar
